requirements.txt
python-telegram-bot>=20.0
requests>=2.28.2
beautifulsoup4>=4.11.1
timezonefinder>=6.1.9
python-telegram-bot[job-queue]>=20.0
lxml>=4.9.2
python-dotenv>=1.0.0